+++ doc/test-suite.py
@@ -50,8 +50,8 @@ if __name__ == '__main__':
                 shortlist.append(line.strip('\n'))
                 print('File {} added to the shortlist'.format(line))
 
-    path1 = args.test_suite_dir.replace("\ ", " ")
-    path2 = args.output_dir.replace("\ ", " ")
+    path1 = args.test_suite_dir.replace(r"\ ", " ")
+    path2 = args.output_dir.replace(r"\ ", " ")
     dir1 = sorted(os.listdir(path1))
     for item1 in dir1:
         if not (os.path.isdir(os.path.join(path1, item1))):