+++ src/iopae.cpp
@@ -61,7 +61,7 @@ PAEOutput::PAEOutput(Doc *doc) : Output(doc) {}
 
 PAEOutput::~PAEOutput() {}
 
-bool PAEOutput::Export(std::string &output)
+std::string PAEOutput::Export()
 {
     m_docScoreDef = true;
     m_mensural = false;
@@ -73,15 +73,13 @@ bool PAEOutput::Export(std::string &output)
     m_currentDots = -1;
     m_grace = false;
 
-    m_doc->GetFirstScoreDef()->SaveObject(this, false);
+    m_doc->GetFirstScoreDef()->SaveObject(this);
 
     m_docScoreDef = false;
 
-    m_doc->SaveObject(this, false);
+    m_doc->SaveObject(this);
 
-    output = m_streamStringOutput.str();
-
-    return true;
+    return m_streamStringOutput.str();
 }
 
 bool PAEOutput::WriteObject(Object *object)
@@ -3277,7 +3275,7 @@ void PAEInput::ParseHeader(jsonxx::Object &header)
     }
 
     bool hasIncip = false;
-    for (const std::string &key : { "scoring", "key_mode", "role", "voice_intrument" }) {
+    for (const char *key : { "scoring", "key_mode", "role", "voice_intrument" }) {
         hasIncip = hasIncip || header.has<jsonxx::String>(key) || header.has<jsonxx::Array>(key);
     }
     if (hasIncip) {