+++ src/savefunctor.cpp
@@ -10,7 +10,7 @@
 //----------------------------------------------------------------------------
 
 #include "editorial.h"
-#include "iomei.h"
+#include "iobase.h"
 #include "mdiv.h"
 #include "mnum.h"
 #include "runningelement.h"
@@ -24,10 +24,11 @@ namespace vrv {
 // SaveFunctor
 //----------------------------------------------------------------------------
 
-SaveFunctor::SaveFunctor(Output *output, bool basic) : Functor()
+SaveFunctor::SaveFunctor(Output *output) : Functor()
 {
+    assert(output);
+
     m_output = output;
-    m_basic = basic;
 }
 
 FunctorCode SaveFunctor::VisitDots(Dots *dots)
@@ -44,23 +45,16 @@ FunctorCode SaveFunctor::VisitDotsEnd(Dots *dots)
 FunctorCode SaveFunctor::VisitEditorialElement(EditorialElement *editorialElement)
 {
     // When writing MEI basic, only visible elements within editorial markup are saved
-    if (m_basic && (editorialElement->IsHidden())) {
-        return FUNCTOR_SIBLINGS;
-    }
-    else {
-        return this->VisitObject(editorialElement);
-    }
+    if (m_output->Skip(editorialElement)) return FUNCTOR_SIBLINGS;
+
+    return this->VisitObject(editorialElement);
 }
 
 FunctorCode SaveFunctor::VisitEditorialElementEnd(EditorialElement *editorialElement)
 {
-    // Same as above
-    if (m_basic && (editorialElement->IsHidden())) {
-        return FUNCTOR_SIBLINGS;
-    }
-    else {
-        return this->VisitObjectEnd(editorialElement);
-    }
+    if (m_output->Skip(editorialElement)) return FUNCTOR_SIBLINGS;
+
+    return this->VisitObjectEnd(editorialElement);
 }
 
 FunctorCode SaveFunctor::VisitFlag(Flag *flag)
@@ -76,23 +70,17 @@ FunctorCode SaveFunctor::VisitFlagEnd(Flag *flag)
 
 FunctorCode SaveFunctor::VisitMdiv(Mdiv *mdiv)
 {
-    MEIOutput *meiOutput = dynamic_cast<MEIOutput *>(m_output);
+    // Can be skipped in MEI output
+    if (m_output->Skip(mdiv)) return FUNCTOR_SIBLINGS;
 
-    if (meiOutput && (mdiv->IsHidden())) {
-        // Do not output hidden mdivs in page-based MEI or when saving score-based MEI with filter
-        if (!meiOutput->GetScoreBasedMEI() || meiOutput->HasFilter()) return FUNCTOR_SIBLINGS;
-    }
     return this->VisitObject(mdiv);
 }
 
 FunctorCode SaveFunctor::VisitMdivEnd(Mdiv *mdiv)
 {
-    MEIOutput *meiOutput = dynamic_cast<MEIOutput *>(m_output);
+    // Can be skipped in MEI output
+    if (m_output->Skip(mdiv)) return FUNCTOR_SIBLINGS;
 
-    if (meiOutput && (mdiv->IsHidden())) {
-        // Do not output hidden mdivs in page-based MEI or when saving score-based MEI with filter
-        if (!meiOutput->GetScoreBasedMEI() || meiOutput->HasFilter()) return FUNCTOR_SIBLINGS;
-    }
     return this->VisitObjectEnd(mdiv);
 }
 
@@ -108,12 +96,18 @@ FunctorCode SaveFunctor::VisitMeasureEnd(Measure *measure)
 
 FunctorCode SaveFunctor::VisitMNum(MNum *mNum)
 {
-    return (mNum->IsGenerated()) ? FUNCTOR_SIBLINGS : this->VisitObject(mNum);
+    // Can be skipped in MEI output
+    if (m_output->Skip(mNum)) return FUNCTOR_SIBLINGS;
+
+    return this->VisitObject(mNum);
 }
 
 FunctorCode SaveFunctor::VisitMNumEnd(MNum *mNum)
 {
-    return (mNum->IsGenerated()) ? FUNCTOR_SIBLINGS : this->VisitObjectEnd(mNum);
+    // Can be skipped in MEI output
+    if (m_output->Skip(mNum)) return FUNCTOR_SIBLINGS;
+
+    return this->VisitObjectEnd(mNum);
 }
 
 FunctorCode SaveFunctor::VisitObject(Object *object)